Time: 5 ms (71.53%), Space: 43.6 MB (91.23%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.java
@@ -1,8 +1,8 @@
 class Solution {
     public boolean isPalindrome(int x) {
         if(x < 0) return false;
-        long reversed = 0;
-        long temp = x;
+        int reversed = 0;
+        int temp = x;
         while(temp != 0){
             int digit = (int)(temp%10);
             reversed = reversed * 10 + digit;
